requirements.txt
altgraph==0.17
cycler==0.10.0
future==0.18.2
graphviz==0.15
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.3
pandas==1.1.4
pefile==2019.4.18
Pillow==8.0.1
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pywin32-ctypes==0.2.0
six==1.15.0
ua-parser==0.10.0
user-agents==2.2.0
